fix(api): moves test from repository to usecase
