+++ api/tests/prescription/campaign/integration/domain/usecases/create-campaign_test.js
@@ -9,7 +9,8 @@ import { CampaignTypes } from '../../../../../../src/prescription/shared/domain/
 import { CAMPAIGN_FEATURES, ORGANIZATION_FEATURE } from '../../../../../../src/shared/domain/constants.js';
 import * as codeGenerator from '../../../../../../src/shared/domain/services/code-generator.js';
 import * as accessCodeRepository from '../../../../../../src/shared/infrastructure/repositories/access-code-repository.js';
-import { databaseBuilder, expect, mockLearningContent } from '../../../../../test-helper.js';
+import { catchErr, databaseBuilder, expect, knex, mockLearningContent } from '../../../../../test-helper.js';
+
 describe('Integration | UseCases | create-campaign', function () {
   let userId;
   let organizationId;
@@ -35,76 +36,6 @@ describe('Integration | UseCases | create-campaign', function () {
     };
 
     await mockLearningContent(learningContent);
-  });
-
-  it('should save a new campaign of type ASSESSMENT', async function () {
-    // given
-    const campaign = {
-      name: 'a name',
-      type: CampaignTypes.ASSESSMENT,
-      title: 'a title',
-      externalIdLabel: 'id Pix label',
-      externalIdType: 'STRING',
-      customLandingPageText: 'Hello',
-      creatorId: userId,
-      ownerId: userId,
-      organizationId,
-      targetProfileId,
-    };
-
-    const expectedAttributes = ['type', 'title', 'externalIdLabel', 'name', 'customLandingPageText'];
-
-    // when
-    const result = await createCampaign({
-      campaign,
-      userRepository,
-      campaignAdministrationRepository,
-      campaignCreatorRepository,
-      codeGenerator,
-      accessCodeRepository,
-    });
-
-    // then
-    expect(result).to.be.an.instanceOf(Campaign);
-
-    expect(_.pick(result, expectedAttributes)).to.deep.equal(_.pick(campaign, expectedAttributes));
-  });
-
-  it('should save a new campaign of type EXAM', async function () {
-    const featureId = databaseBuilder.factory.buildFeature(ORGANIZATION_FEATURE.CAMPAIGN_WITHOUT_USER_PROFILE).id;
-
-    databaseBuilder.factory.buildOrganizationFeature({ featureId, organizationId });
-    await databaseBuilder.commit();
-    // given
-    const campaign = {
-      name: 'a name',
-      type: CampaignTypes.EXAM,
-      title: 'a title',
-      externalIdLabel: 'id Pix label',
-      externalIdType: 'STRING',
-      customLandingPageText: 'Hello',
-      creatorId: userId,
-      ownerId: userId,
-      organizationId,
-      targetProfileId,
-    };
-
-    const expectedAttributes = ['type', 'title', 'externalIdLabel', 'name', 'customLandingPageText'];
-
-    // when
-    const result = await createCampaign({
-      campaign,
-      userRepository,
-      campaignAdministrationRepository,
-      campaignCreatorRepository,
-      codeGenerator,
-      accessCodeRepository,
-    });
-
-    // then
-    expect(result).to.be.an.instanceOf(Campaign);
-
-    expect(_.pick(result, expectedAttributes)).to.deep.equal(_.pick(campaign, expectedAttributes));
   });
 
   it('should save a new campaign of type PROFILES_COLLECTION', async function () {
@@ -134,5 +65,196 @@ describe('Integration | UseCases | create-campaign', function () {
     expect(result).to.be.an.instanceOf(Campaign);
     expect(_.pick(result, expectedAttributes)).to.deep.equal(_.pick(campaign, expectedAttributes));
     expect(result.code).to.have.lengthOf.above(0);
+  });
+  describe('type ASSESSMENT', function () {
+    it('should not save anything if something goes wrong between campaign creation and skills computation', async function () {
+      // given
+      const learningContent = {
+        areas: [{ id: 'recArea1', competenceIds: ['recCompetence1'] }],
+        competences: [
+          {
+            id: 'recCompetence1',
+            areaId: 'recArea1',
+            tubeIds: ['recTube1', 'recTube2', 'recTube3'],
+          },
+        ],
+        tubes: [
+          {
+            id: 'recTube1',
+            skillIds: ['recSkill2'],
+          },
+        ],
+        skills: [
+          {
+            id: 'recSkill2',
+            name: 'recSkill2',
+            status: 'actif',
+            level: 1,
+            tubeId: 'recTube1',
+          },
+        ],
+      };
+      await mockLearningContent(learningContent);
+
+      databaseBuilder.factory.buildTargetProfileTube({ targetProfileId, tubeId: 'recTube1', level: 2 });
+      await databaseBuilder.commit();
+      const campaignToSave = {
+        name: 'Evaluation niveau 1 recherche internet',
+        code: 'BCTERD153',
+        customLandingPageText: 'Parcours évaluatif concernant la recherche internet',
+        creatorId: userId,
+        ownerId: userId,
+        organizationId,
+        multipleSendings: true,
+        type: CampaignTypes.ASSESSMENT,
+        targetProfileId,
+        title: 'Parcours recherche internet',
+      };
+
+      // when
+      await catchErr(createCampaign)(campaignToSave, {
+        userRepository,
+        campaignAdministrationRepository,
+        campaignCreatorRepository,
+        codeGenerator,
+        accessCodeRepository,
+      });
+
+      // then
+      const skillIds = await knex('campaign_skills').pluck('skillId');
+      const campaignIds = await knex('campaigns').pluck('id');
+      expect(skillIds).to.be.empty;
+      expect(campaignIds).to.be.empty;
+    });
+    it('should save a new campaign of type ASSESSMENT', async function () {
+      // given
+      const campaign = {
+        name: 'a name',
+        type: CampaignTypes.ASSESSMENT,
+        title: 'a title',
+        externalIdLabel: 'id Pix label',
+        externalIdType: 'STRING',
+        customLandingPageText: 'Hello',
+        creatorId: userId,
+        ownerId: userId,
+        organizationId,
+        targetProfileId,
+      };
+
+      const expectedAttributes = ['type', 'title', 'externalIdLabel', 'name', 'customLandingPageText'];
+
+      // when
+      const result = await createCampaign({
+        campaign,
+        userRepository,
+        campaignAdministrationRepository,
+        campaignCreatorRepository,
+        codeGenerator,
+        accessCodeRepository,
+      });
+
+      // then
+      expect(result).to.be.an.instanceOf(Campaign);
+
+      expect(_.pick(result, expectedAttributes)).to.deep.equal(_.pick(campaign, expectedAttributes));
+    });
+  });
+  describe('type EXAM', function () {
+    it('should not save anything if something goes wrong between campaign creation and skills computation', async function () {
+      // given
+      const learningContent = {
+        areas: [{ id: 'recArea1', competenceIds: ['recCompetence1'] }],
+        competences: [
+          {
+            id: 'recCompetence1',
+            areaId: 'recArea1',
+            tubeIds: ['recTube1', 'recTube2', 'recTube3'],
+          },
+        ],
+        tubes: [
+          {
+            id: 'recTube1',
+            skillIds: ['recSkill3'],
+          },
+        ],
+        skills: [
+          {
+            id: 'recSkill3',
+            name: 'recSkill3',
+            status: 'actif',
+            level: 1,
+            tubeId: 'recTube1',
+          },
+        ],
+      };
+      await mockLearningContent(learningContent);
+
+      databaseBuilder.factory.buildTargetProfileTube({ targetProfileId, tubeId: 'recTube1', level: 2 });
+      await databaseBuilder.commit();
+      const campaignToSave = {
+        name: 'Evaluation niveau 1 recherche internet',
+        code: 'BCTERD153',
+        customLandingPageText: 'Parcours évaluatif concernant la recherche internet',
+        creatorId: userId,
+        ownerId: userId,
+        organizationId,
+        multipleSendings: true,
+        type: CampaignTypes.EXAM,
+        targetProfileId,
+        title: 'Parcours recherche internet',
+      };
+
+      // when
+      // when
+      await catchErr(createCampaign)(campaignToSave, {
+        userRepository,
+        campaignAdministrationRepository,
+        campaignCreatorRepository,
+        codeGenerator,
+        accessCodeRepository,
+      });
+
+      // then
+      const skillIds = await knex('campaign_skills').pluck('skillId');
+      const campaignIds = await knex('campaigns').pluck('id');
+      expect(skillIds).to.be.empty;
+      expect(campaignIds).to.be.empty;
+    });
+    it('should save a new campaign of type EXAM', async function () {
+      const featureId = databaseBuilder.factory.buildFeature(ORGANIZATION_FEATURE.CAMPAIGN_WITHOUT_USER_PROFILE).id;
+
+      databaseBuilder.factory.buildOrganizationFeature({ featureId, organizationId });
+      await databaseBuilder.commit();
+      // given
+      const campaign = {
+        name: 'a name',
+        type: CampaignTypes.EXAM,
+        title: 'a title',
+        externalIdLabel: 'id Pix label',
+        externalIdType: 'STRING',
+        customLandingPageText: 'Hello',
+        creatorId: userId,
+        ownerId: userId,
+        organizationId,
+        targetProfileId,
+      };
+
+      const expectedAttributes = ['type', 'title', 'externalIdLabel', 'name', 'customLandingPageText'];
+
+      // when
+      const result = await createCampaign({
+        campaign,
+        userRepository,
+        campaignAdministrationRepository,
+        campaignCreatorRepository,
+        codeGenerator,
+        accessCodeRepository,
+      });
+
+      // then
+      expect(result).to.be.an.instanceOf(Campaign);
+
+      expect(_.pick(result, expectedAttributes)).to.deep.equal(_.pick(campaign, expectedAttributes));
+    });
   });
 });
